fix: fixed migration order

+++ database/migrations/0001_01_01_000000_create_users_table.php
@@ -11,17 +11,19 @@ return new class extends Migration
      */
     public function up(): void
     {
-        Schema::create('users', function (Blueprint $table) {
-            $table->id();
-            $table->string('name');
-            $table->string('email')->unique();
-            $table->timestamp('email_verified_at')->nullable();
-            $table->string('password');
-            $table->enum('role', ['admin', 'manager', 'scanner'])->default('scanner');
-            $table->foreignId('department_id')->nullable()->constrained()->onDelete('set null');
-            $table->rememberToken();
-            $table->timestamps();
-        });
+        if (!Schema::hasTable('users')) {
+            Schema::create('users', function (Blueprint $table) {
+                $table->id();
+                $table->string('name');
+                $table->string('email')->unique();
+                $table->timestamp('email_verified_at')->nullable();
+                $table->string('password');
+                $table->enum('role', ['admin', 'manager', 'scanner'])->default('scanner');
+                $table->foreignId('department_id')->nullable()->constrained()->onDelete('set null');
+                $table->rememberToken();
+                $table->timestamps();
+            });
+        }
     }
 
     /**
